refactor: make Pagination total a long

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/pagination/PaginationInfo.java
@@ -22,7 +22,7 @@ public class PaginationInfo {
 
     private PaginationInfo() {}
 
-    public static Pagination computePaginationInfo(Integer totalCount, Integer pageItemsCount, PaginationParam paginationParam) {
+    public static Pagination computePaginationInfo(long totalCount, Integer pageItemsCount, PaginationParam paginationParam) {
         Pagination pagination = new Pagination();
         if (totalCount > 0) {
             pagination
