+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/pagination/PaginationLinks.java
@@ -27,7 +27,7 @@ public class PaginationLinks {
     public static Links computePaginationLinks(
         URI requestUri,
         MultivaluedMap<String, String> queryParameters,
-        int totalElements,
+        long totalElements,
         PaginationParam paginationParam
     ) {
         if (totalElements == 0) {
